requirements.txt
# Core agent & memory (adaptează versiuni dacă ai recomandări de la vendor)
agno>=0.1.0
mem0ai>=0.1.0

# LLM provider client
openai>=0.27.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0

# For trends/plots (opțional)
pandas>=2.0.0
matplotlib>=3.7.0
textblob>=0.19.0

